Improved caching in OTRSCodePolicy unit test.
We do this by getting a cache version MD5 string that changes when the OTRSCodePolicy module changes.
This is done by getting all file names and contents in Kernel/TidyAll and computing an MD5 on it, and will allow us to cache for a long time (30 days).

Notify: marc.bonsels@otrs.com, manuel.hecht@otrs.com